EPC calc, Q_a added, Reorganized Data Exchange
There was an error in the location of the EPC calculation. EPC used
DNIatModule, which was calculated after the EPC equation. This should
not have an effect on the results because EPC is not directly used.

Added Q_a equation but need to consider the validity of it based on the
assumption in the Q_c equation. Q_c accounts for the convertion
efficiencies. That which is not converted is heat so it should account
of all energy losses...but does it.

Updated how the data is passed ot the getRadiativeGain equation. There
were far too many variables being passed, easier to pass the dictionary
and reference the information.

Next is adding logic that considers EPC. There are times when the EPC is
0 because the tracking angles are beyond the systems reach but there
would still be solar radiation coming into the building facade that
needs to be considered in the heat gain calculation.