Fixed clockwise perspective error when reading IDF geometry
When writing the code for vertices the perspective was from inside the
building looking outwards. EnergyPlus parameter for "Vertex Entry
Direct" with values "counterclockwise" and "clockwise" are written from
outside the building looking inwards perspective. This required
reversing the vertices to match that of the direction in which the
python script was written by Dan.